Implement partition header, update .gitignore
